process-related promises via '/proc'
Process-related promises have hitherto been implemented by invoking the
'ps' command and attempting to parse its output.  But 'ps' output
is highly human-oriented, and often not easy to machine-parse reliably.
It is also highly OS-variable.

CFE-1655 had initially proposed the idea of using "/proc" where available.
My own encounter with the flaws in the existing mechanism is at CFE-3020.

Ideally, this code would be chosen whenever "/proc" is available, and an
autotools "HAVE_PS_VIA_PROC" variable is used to mange this opt-in.

In this first implementation (commit), I have limited it to Linux
(my development platforms have been RHEL6 and RHEL7) and this extra
check, which would ideally be removed in due course, is governed
by a clause in "configure.ac".

This first commit passes most "tests/acceptance/05_processes/01_matching"
tests, including two recently added ones.  It currently fails two
existing ones, so needs a little further work.

There may also be a little more work needed to cover the case of
a process exiting during the small but finite period while the
information about it is being gathered.