unit tests: bring new '/proc' method up to same level as 'ps' method
The unit tests for process handling now offer the same coverage
for the new '/proc' as they had done for the traditional 'ps'.

The missing link had been the "linux_process_test".
This now goes through the new code rather than the older "process_linux.c",
which is now redundant.
Because this test is now generic "/proc" rather than solely Linux,
it is renamed to "proc_process_test".