process handling: rename control mechanism
This project had started out to use "/proc" to replace 'ps' output for handling processes
where this is possible.  This is fine for Linux and (we believe) for Solaris.
But we then observe that the BSD family don't have "/proc", yet does offer alternative
mechanisms, such as "kvm_getprocs()".

So we need to reframe our thinking from "/proc is better" to "probing the OS is better".
This commit renames files, autotools variable and CPP variables from "/proc/" to "/sysprobe/".
It also rephrases some associated comments.