process handling: partially revert a much earlier change
BSD development is shaping up well.  But it revealed an earlier poor assumption.
We had hitherto assumed that the way to avoid reliance on ps-parsing
was to use "/proc".  But BSD doesn't offer "/proc", yet does offer
the "kvm_getprocs(3)" mechanism.  So we need to adjust the model from
"/proc replaces ps" to "an OS-appropriate interface replaces ps".
The "tests/unit/" shift towards "/proc" was thus a step too far,
and we revert it here.